+++ src/Data/Entity/Notification/NotificationDefinition.php
@@ -1,0 +1,51 @@
+<?php declare(strict_types=1);
+namespace Vin\ShopwareSdk\Data\Entity\Notification;
+
+use Vin\ShopwareSdk\Data\Entity\EntityDefinition;
+use Vin\ShopwareSdk\Data\Schema\PropertyCollection;
+use Vin\ShopwareSdk\Data\Schema\FlagCollection;
+use Vin\ShopwareSdk\Data\Schema\Property;
+use Vin\ShopwareSdk\Data\Schema\Flag;
+use Vin\ShopwareSdk\Data\Schema\Schema;
+
+/**
+ * Shopware Definition Mapping Class
+ *
+ * This class is generated dynamically following SW entities schema
+ */
+class NotificationDefinition implements EntityDefinition
+{
+    public const ENTITY_NAME = 'notification';
+
+    public function getEntityName() : string
+    {
+        return self::ENTITY_NAME;
+    }
+
+    public function getEntityClass() : string
+    {
+        return NotificationEntity::class;
+    }
+
+    public function getEntityCollection() : string
+    {
+        return NotificationCollection::class;
+    }
+
+    public function getSchema() : Schema
+    {
+        return new Schema('notification', new PropertyCollection([
+            new Property('id', 'uuid', new FlagCollection([new Flag('read_protected', [['Shopware\Core\Framework\Api\Context\AdminApiSource']]), new Flag('primary_key', 1), new Flag('required', 1), ]), []),
+            new Property('status', 'string', new FlagCollection([new Flag('read_protected', [['Shopware\Core\Framework\Api\Context\AdminApiSource']]), new Flag('required', 1), ]), []),
+            new Property('message', 'string', new FlagCollection([new Flag('read_protected', [['Shopware\Core\Framework\Api\Context\AdminApiSource']]), new Flag('required', 1), ]), []),
+            new Property('adminOnly', 'boolean', new FlagCollection([new Flag('read_protected', [['Shopware\Core\Framework\Api\Context\AdminApiSource']]), ]), []),
+            new Property('requiredPrivileges', 'json_list', new FlagCollection([new Flag('read_protected', [['Shopware\Core\Framework\Api\Context\AdminApiSource']]), ]), []),
+            new Property('createdByIntegrationId', 'uuid', new FlagCollection([new Flag('read_protected', [['Shopware\Core\Framework\Api\Context\AdminApiSource']]), ]), []),
+            new Property('createdByUserId', 'uuid', new FlagCollection([new Flag('read_protected', [['Shopware\Core\Framework\Api\Context\AdminApiSource']]), ]), []),
+            new Property('createdByIntegration', 'association', new FlagCollection([new Flag('read_protected', [['Shopware\Core\Framework\Api\Context\AdminApiSource']]), ]), ['entity' => 'integration', 'referenceField' => 'id', 'localField' => 'createdByIntegrationId', 'relation' => 'many_to_one', ]),
+            new Property('createdByUser', 'association', new FlagCollection([new Flag('read_protected', [['Shopware\Core\Framework\Api\Context\AdminApiSource']]), ]), ['entity' => 'user', 'referenceField' => 'id', 'localField' => 'createdByUserId', 'relation' => 'many_to_one', ]),
+            new Property('createdAt', 'date', new FlagCollection([new Flag('read_protected', [['Shopware\Core\Framework\Api\Context\AdminApiSource', 'Shopware\Core\Framework\Api\Context\SalesChannelApiSource']]), new Flag('required', 1), ]), []),
+            new Property('updatedAt', 'date', new FlagCollection([new Flag('read_protected', [['Shopware\Core\Framework\Api\Context\AdminApiSource', 'Shopware\Core\Framework\Api\Context\SalesChannelApiSource']]), ]), []),
+        ]));
+    }
+}