+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to `shopware-php-sdk` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+### 1.4.0
+- Updated Latest DAL Classes
+- Added NotificationService to allow sending/getting notification from external app (success/error/warning/info)
+- Added UserConfigService to getting/saving admin's config
+- Added AdminSearchService to search for multiple entities as one request
+- Added more service's examples
+- Deprecated SystemConfigService::saveConfiguration, use SystemConfigService::save instead
+- Deprecated SystemConfigService::batchSaveConfiguration, use SystemConfigService::batchSave instead
+
 ### 1.3.3
 - [Use correct sync operator for syncDeleted](https://github.com/vienthuong/shopware-php-sdk/pull/16)
 - [Remove extends HttpClient from GuzzleClient](https://github.com/vienthuong/shopware-php-sdk/issues/5)